fix task duration timestamp

+++ models/taskModel.js
@@ -2,7 +2,15 @@ const { pool } = require("../server");
 
 async function getTasks(userId) {
   const result = await pool.query(
-    "SELECT * FROM tasks WHERE user_id = $1 ORDER BY created_at ASC;",
+    `SELECT 
+      id, 
+      name, 
+      status, 
+      created_at, 
+      user_id, 
+      TO_CHAR(duration, 'HH24:MI') as duration, 
+      start_time 
+    FROM tasks WHERE user_id = $1 ORDER BY created_at ASC;`,
     [userId]
   );
   return result.rows;
@@ -17,10 +25,10 @@ async function getTaskById(id) {
 }
 
 async function createTask(taskData) {
-  const { name, status, user_id, duration, start_time } = taskData;
+  const { name, user_id, duration, start_time } = taskData;
   const result = await pool.query(
-    "INSERT INTO tasks (name, status, user_id, duration, start_time) VALUES ($1, $2, $3, $4, $5) RETURNING *;",
-    [name, status, user_id, duration, start_time]
+    "INSERT INTO tasks (name, user_id, duration, start_time) VALUES ($1, $2, ($3 || ':00')::interval, $4) RETURNING id, name, status, created_at, user_id, TO_CHAR(duration, 'HH24:MI') as duration, start_time;",
+    [name, user_id, duration, start_time]
   );
   return result.rows[0];
 }
@@ -30,15 +38,29 @@ async function updateTask(id, taskData) {
 
   const fieldsToUpdate = Object.keys(fields);
 
+  const durationIndex = fieldsToUpdate.indexOf("duration");
+  if (durationIndex !== -1) {
+    fieldsToUpdate[durationIndex] = `($${
+      durationIndex + 1
+    } || ':00')::interval`;
+  }
+
   const fieldsString = fieldsToUpdate
-    .map((field, index) => `${field} = $${index + 1}`)
+    .map((field, index) => {
+      if (field.includes("::interval")) {
+        return `duration = ${field}`;
+      }
+      return `${field} = $${index + 1}`;
+    })
     .join(", ");
 
-  const values = fieldsToUpdate.map((item) => fields[item]);
+  const values = Object.keys(fields).map((item) => fields[item]);
 
   const query = `UPDATE tasks SET ${fieldsString} WHERE id = $${
-    fieldsToUpdate.length + 1
-  } AND user_id = $${fieldsToUpdate.length + 2} RETURNING *;`;
+    values.length + 1
+  } AND user_id = $${
+    values.length + 2
+  } RETURNING id, name, status, created_at, user_id, TO_CHAR(duration, 'HH24:MI') as duration, start_time;`;
 
   const result = await pool.query(query, [...values, id, user_id]);
   return result.rows[0];
@@ -46,7 +68,7 @@ async function updateTask(id, taskData) {
 
 async function deleteTaskById(id, userId) {
   const result = await pool.query(
-    "DELETE FROM tasks WHERE id = $1 AND user_id = $2 RETURNING *;",
+    "DELETE FROM tasks WHERE id = $1 AND user_id = $2 RETURNING id, name, status, created_at, user_id, TO_CHAR(duration, 'HH24:MI') as duration, start_time;",
     [id, userId]
   );
   return result.rows[0];
